requirements.txt
streamlit==1.18.0
google-cloud-aiplatform==1.26.0
asyncio==3.4.3
asyncpg==0.27.0 
numpy==1.22.4 
pandas==1.5.3
pgvector==0.1.8
langchain==0.0.196 
transformers==4.30.1
bs4
torch
altair<5
protobuf<=3.20.1
cloud-sql-python-connector
faiss-cpu
lxml
fake-useragent